refactor(backends): move openshell into midojo.backends subpackage
Convert backends.py into a backends/ package and relocate the OpenShell
backend and its OCSF log helper into it:

  backends.py          -> backends/__init__.py
  openshell_backend.py -> backends/openshell.py
  openshell_logs.py    -> backends/openshell_logs.py

Update imports in agent_client, orchestrator, and the openshell tests.
The public midojo.backends API (EnvironmentBackend, build_backend,
register_backend, DictEnvironmentBackend) is unchanged; only the
openshell module path moves to midojo.backends.openshell.
